VM: huge pages for HHDM + refactoring.
Implemented function `resolve_vaddr` that  takes care of the hint and flags provided to the map function.

+++ kernel/include/arch/types.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <stdint.h>
+#include <stddef.h>
 
 #if defined(__x86_64__)
 
@@ -11,7 +12,7 @@
 #define ARCH_PAGE_SIZE_1G 0x40000000ull
 #define ARCH_PAGE_GRAN ARCH_PAGE_SIZE_4K
 
-#define ARCH_PAGE_SIZES ((u64[]){ARCH_PAGE_SIZE_4K, ARCH_PAGE_SIZE_2M, ARCH_PAGE_SIZE_1G})
+#define ARCH_PAGE_SIZES ((size_t[]){ARCH_PAGE_SIZE_4K, ARCH_PAGE_SIZE_2M, ARCH_PAGE_SIZE_1G})
 #define ARCH_PAGE_SIZES_LEN 3
 
 #elif defined(__aarch64__)
@@ -23,7 +24,7 @@
 #define ARCH_PAGE_SIZE_1G 0x40000000ull
 #define ARCH_PAGE_GRAN ARCH_PAGE_SIZE_4K
 
-#define ARCH_PAGE_SIZES ((u64[]){ARCH_PAGE_SIZE_4K, ARCH_PAGE_SIZE_2M, ARCH_PAGE_SIZE_1G})
+#define ARCH_PAGE_SIZES ((size_t[]){ARCH_PAGE_SIZE_4K, ARCH_PAGE_SIZE_2M, ARCH_PAGE_SIZE_1G})
 #define ARCH_PAGE_SIZES_LEN 3
 
 #endif
